big refactor

+++ One/WeatherForecast.cs
@@ -4,8 +4,8 @@ namespace One
 {
     public class WeatherForecast
     {
-        public double Lat { get; set; }
-        public double Lon { get; set; }
+        public decimal Lat { get; set; }
+        public decimal Lon { get; set; }
 
         public DateTime Date { get; set; }
